remove port 5959, as IPC is used by default
via

```
Christien Rioux
@crioux

This won't work.

EXPOSE 5959 is dangerous, IPC is used by default these days. veilid-cli should just be run as an exec into the docker image.

Dockerfiles to build images, if they are well-vetted should be checked into the veilid repository not the documentation.

This is should not be merged.

```